requirements.txt
cloudscraper==1.2.71
eth-account==0.10.0
web3==6.11.3
pycryptodome==3.19.0